make chainId available to translations
